requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
openpyxl==3.1.2